refactoring, folder for interfaces, folder for functions
